Fix possible NaN return value due to rounding error
Fix an issue in mpc_follower where rounding error in floating point
arithmetic operation could cause a negative number to be passed to
std::sqrt, resulting in an NaN return value. This problem shows up when
fast math optimization is turned on in the compiler.

The issue is fixed by checking the result of the arithmetic operation
and overwriting any negative values with zero. This is safe as the
underlying mathematical value is never expected to be less than zero.

Change-Id: I226e994ee1574871fd978ce6e73fb5f4ae4a471f
Signed-off-by: Liyou Zhou <liyou.zhou@arm.com>